Extract shared AES-256-GCM at-rest crypto helpers
Consolidate the file backend's envelope encrypt/decrypt and wrapping-key
management into a shared util/at-rest module so every at-rest write uses one
authenticated-cipher implementation. No behavior change.
