+++ packages/vaultkeeper/src/backend/file-backend.ts
@@ -14,15 +14,12 @@
 import * as fs from 'node:fs/promises'
 import * as path from 'node:path'
 import * as os from 'node:os'
-import * as crypto from 'node:crypto'
 import { SecretNotFoundError, FilesystemError } from '../errors.js'
+import { encryptGcm, decryptGcm, getOrCreateWrapKey } from '../util/at-rest.js'
 import type { ListableBackend } from './types.js'
 
 const STORAGE_DIR_NAME = path.join('.vaultkeeper', 'file')
 const KEY_FILE = '.key'
-const GCM_IV_BYTES = 12
-const GCM_KEY_BYTES = 32
-const GCM_TAG_LENGTH = 128 // bits
 
 /**
  * Resolve the storage directory for a FileBackend instance.
@@ -60,50 +57,7 @@ async function ensureStorageDir(storageDir: string): Promise<void> {
 }
 
 async function getOrCreateKey(storageDir: string): Promise<Buffer> {
-  const keyPath = path.join(storageDir, KEY_FILE)
-  try {
-    const data = await fs.readFile(keyPath)
-    return data
-  } catch (err) {
-    if (err instanceof Error && 'code' in err && err.code === 'ENOENT') {
-      // Generate a random 32-byte key
-      const key = crypto.randomBytes(GCM_KEY_BYTES)
-      await fs.writeFile(keyPath, key, { mode: 0o600 })
-      return key
-    }
-    throw err
-  }
-}
-
-function encryptGcm(key: Buffer, plaintext: string): string {
-  const iv = crypto.randomBytes(GCM_IV_BYTES)
-  const cipher = crypto.createCipheriv('aes-256-gcm', key, iv, {
-    authTagLength: GCM_TAG_LENGTH / 8,
-  })
-  const encrypted = Buffer.concat([cipher.update(plaintext, 'utf8'), cipher.final()])
-  const authTag = cipher.getAuthTag()
-  return [iv.toString('base64'), authTag.toString('base64'), encrypted.toString('base64')].join(':')
-}
-
-function decryptGcm(key: Buffer, encoded: string): string {
-  const parts = encoded.split(':')
-  if (parts.length !== 3) {
-    throw new Error('Invalid encrypted file format: expected iv:authTag:ciphertext')
-  }
-  const [ivB64, authTagB64, ciphertextB64] = parts
-  if (ivB64 === undefined || authTagB64 === undefined || ciphertextB64 === undefined) {
-    throw new Error('Invalid encrypted file format: missing part')
-  }
-  const iv = Buffer.from(ivB64, 'base64')
-  const authTag = Buffer.from(authTagB64, 'base64')
-  const ciphertext = Buffer.from(ciphertextB64, 'base64')
-
-  const decipher = crypto.createDecipheriv('aes-256-gcm', key, iv, {
-    authTagLength: GCM_TAG_LENGTH / 8,
-  })
-  decipher.setAuthTag(authTag)
-  const decrypted = Buffer.concat([decipher.update(ciphertext), decipher.final()])
-  return decrypted.toString('utf8')
+  return getOrCreateWrapKey(path.join(storageDir, KEY_FILE))
 }
 
 /**
